+++ hutool-core/src/main/java/cn/hutool/core/util/NumberUtil.java
@@ -2137,9 +2137,6 @@ public class NumberUtil {
 		if (StrUtil.startWithIgnoreCase(number, "0x")) {
 			// 0x04表示16进制数
 			return Integer.parseInt(number.substring(2), 16);
-		} else if (number.startsWith("0") && number.length() > 1) {
-			// 04表示8进制数
-			return Integer.parseInt(number.substring(1), 8);
 		}
 
 		return Integer.parseInt(removeNumberFlag(number));
@@ -2173,10 +2170,8 @@ public class NumberUtil {
 		if (number.startsWith("0x")) {
 			// 0x04表示16进制数
 			return Long.parseLong(number.substring(2), 16);
-		} else if (number.startsWith("0") && number.length() > 1) {
-			// 04表示8进制数
-			return Long.parseLong(number.substring(1), 8);
 		}
+		
 		return Long.parseLong(removeNumberFlag(number));
 	}