+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 【core】        增加MurmurHash（Murmur3算法实现），HashUtil增加murmur32、murmur64、murmur128方法
 * 【core】        增加Simhash（用于海量文本去重）
 * 【extra】        增加分词封装，封装了ansj、HanLP、IKAnalyzer、Jcseg、Jieba、MMSeg的实现，统一了接口
+* 【core】        去除NumberUtil.parseInt和parseLong的8进制支持（issue#234@Github）
 
 ### Bug修复